pyproject.toml
[project]
name = "toon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "python-toon>=0.1.3",
]
